templates/database-app/1.0.0/main.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
数据库查询应用主程序
"""

import yaml
from typing import Dict, List, Any, Optional


class DatabaseApp:
    """数据库查询应用类"""
    
    def __init__(self, config_path: str = "config.yaml"):
        """初始化数据库应用"""
        self.config = self.load_config(config_path)
        self.connection = None
    
    def load_config(self, config_path: str) -> Dict:
        """加载配置文件"""
        try:
            with open(config_path, 'r', encoding='utf-8') as f:
                return yaml.safe_load(f)
        except Exception as e:
            raise Exception(f"配置文件加载失败: {e}")
    
    def connect(self) -> bool:
        """建立数据库连接"""
        try:
            db_config = self.config.get('database', {})
            db_type = db_config.get('type', 'mongodb')
            
            # TODO: 根据数据库类型建立连接
            if db_type == 'mongodb':
                self._connect_mongodb(db_config)
            elif db_type == 'mysql':
                self._connect_mysql(db_config)
            elif db_type == 'postgresql':
                self._connect_postgresql(db_config)
            else:
                raise Exception(f"不支持的数据库类型: {db_type}")
            
            return True
        except Exception as e:
            raise Exception(f"数据库连接失败: {e}")
    
    def _connect_mongodb(self, config: Dict):
        """连接 MongoDB"""
        # TODO: 实现 MongoDB 连接
        print(f"连接到 MongoDB: {config.get('host')}:{config.get('port')}")
    
    def _connect_mysql(self, config: Dict):
        """连接 MySQL"""
        # TODO: 实现 MySQL 连接
        print(f"连接到 MySQL: {config.get('host')}:{config.get('port')}")
    
    def _connect_postgresql(self, config: Dict):
        """连接 PostgreSQL"""
        # TODO: 实现 PostgreSQL 连接
        print(f"连接到 PostgreSQL: {config.get('host')}:{config.get('port')}")
    
    def query(self, query_str: str, params: Optional[Dict] = None) -> Dict[str, Any]:
        """执行查询"""
        try:
            if not self.connection:
                self.connect()
            
            # TODO: 实现具体的查询逻辑
            results = self._execute_query(query_str, params or {})
            
            return {
                "status": "success",
                "data": results,
                "count": len(results)
            }
        except Exception as e:
            return {
                "status": "error",
                "message": str(e)
            }
    
    def _execute_query(self, query_str: str, params: Dict) -> List[Dict]:
        """执行实际查询操作"""
        # TODO: 根据数据库类型执行查询
        # 这里是示例代码，需要根据实际需求修改
        
        results = []
        # 示例查询结果
        for i in range(3):
            results.append({
                "id": i + 1,
                "name": f"记录 {i + 1}",
                "query": query_str
            })
        
        return results
    
    def close(self):
        """关闭数据库连接"""
        if self.connection:
            self.connection.close()
            self.connection = None


def main():
    """主函数"""
    app = DatabaseApp()
    
    try:
        # 示例查询
        result = app.query("SELECT * FROM users LIMIT 5")
        print(result)
    finally:
        app.close()


if __name__ == "__main__":
    main()