templates/database-app/requirements.txt
pyyaml>=6.0
pymongo>=4.3.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=1.4.0